+++ src/main/java/br/com/devschool/devschool/service/Resposta/RespostaServiceImpl.java
@@ -1,10 +1,10 @@
 package br.com.devschool.devschool.service.Resposta;
 
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.stereotype.Service;
 
-import br.com.devschool.devschool.infrastructure.exception.ContentNotFoundException;
 import br.com.devschool.devschool.model.Disciplina;
 import br.com.devschool.devschool.model.Resposta;
 import br.com.devschool.devschool.model.formDto.RespostaFormDTO;
@@ -21,27 +21,22 @@ public class RespostaServiceImpl implements RespostaService {
 
 	@Override
 	public List<Resposta> listarRespostas(Integer disciplinaId) {
-		if(disciplinaId != null){
-			return respostaRepository.findAllByDisciplinaId(disciplinaId);
-		}else{
-			return respostaRepository.findAll();
-		}
-
-	}
-
-	@Override
-	public Resposta listarRespostaById(Integer id) {
-		return respostaRepository.findById(id)
-				.orElseThrow(() -> new ContentNotFoundException("Resposta com id: " + id + "não encontrada"));
+		return respostaRepository.findAllByDisciplinaId(disciplinaId);
 	}
 
 	@Override
 	public Resposta inserirResposta(RespostaFormDTO respostaDTO) {
-		Disciplina disciplina = disciplinaRepository.findById(respostaDTO.getDisciplinaId())
-				.orElseThrow(() -> new ContentNotFoundException("Disciplina com id "+ respostaDTO.getDisciplinaId() +" não encontrada"));
+		Disciplina disciplina = null;
+		if (respostaDTO.getDisciplinaId() != null) {
+			Optional<Disciplina> disciplinaOptional = disciplinaRepository.findById(respostaDTO.getDisciplinaId());
+			if (disciplinaOptional.isEmpty()) {
+				throw new RuntimeException("Disciplina inexistente");
+			}
+			disciplina = disciplinaOptional.get();
+		}
 		
 		Resposta resposta = Resposta.builder()
-				.conteudo(respostaDTO.getConteudo())
+				.descricao(respostaDTO.getDescricao())
 				.disciplina(disciplina)
 				.build();
 		return respostaRepository.save(resposta);
@@ -49,23 +44,33 @@ public class RespostaServiceImpl implements RespostaService {
 
 	@Override
 	public Resposta alterarResposta(Integer id, RespostaFormDTO respostaDTO) {
-		Resposta resposta = respostaRepository.findById(id)
-				.orElseThrow(() -> new ContentNotFoundException("Resposta com id "+ id +" não encontrada"));
+		Optional<Resposta> respostaOptional = respostaRepository.findById(id);
+		if (respostaOptional.isEmpty()) {
+			throw new RuntimeException("Resposta não encontrada");
+		}
+		Resposta resposta = respostaOptional.get();
 
-		Disciplina disciplina = disciplinaRepository.findById(respostaDTO.getDisciplinaId())
-				.orElseThrow(() -> new ContentNotFoundException("Disciplina com id "+ respostaDTO.getDisciplinaId() +" não encontrada"));
+		Disciplina disciplina = null;
+		if (respostaDTO.getDisciplinaId() != null) {
+			Optional<Disciplina> disciplinaOptional = disciplinaRepository.findById(respostaDTO.getDisciplinaId());
+			if (disciplinaOptional.isEmpty()) {
+				throw new RuntimeException("Disciplina inexistente");
+			}
+			disciplina = disciplinaOptional.get();
+		}
 		
-		resposta.setConteudo(respostaDTO.getConteudo());
+		resposta.setDescricao(respostaDTO.getDescricao());
 		resposta.setDisciplina(disciplina);
-		
 		return respostaRepository.save(resposta);
 	}
 
 	@Override
 	public void excluirResposta(Integer id) {
-		Resposta resposta = respostaRepository.findById(id)
-				.orElseThrow(() -> new ContentNotFoundException("Resposta com id " + id + " não encontrada"));
-		
+		Optional<Resposta> respostaOptional = respostaRepository.findById(id);
+		if (respostaOptional.isEmpty()) {
+			throw new RuntimeException("Resposta não encontrada");
+		}
+
 		respostaRepository.deleteById(id);
 	}
 }
